No more get4

+++ n64-sys/src/sys.rs
@@ -3,7 +3,7 @@ use core::{arch::asm, mem::size_of};
 #[inline]
 pub unsafe fn data_cache_hit_writeback_invalidate<T>(block: &[T]) {
     let addr = (block.as_ptr() as usize) & 0xffff_fff0;
-    let len = block.len() * size_of::<T>() + (block.as_ptr() as usize - addr as usize);
+    let len = block.len() * size_of::<T>() + (block.as_ptr() as usize - addr);
     let mut i = 0;
 
     while i < len {
@@ -25,7 +25,7 @@ pub unsafe fn data_cache_hit_writeback_invalidate_single(addr: usize) {
 #[inline]
 pub unsafe fn data_cache_hit_writeback<T>(block: &[T]) {
     let addr = (block.as_ptr() as usize) & 0xffff_fff0;
-    let len = block.len() * size_of::<T>() + (block.as_ptr() as usize - addr as usize);
+    let len = block.len() * size_of::<T>() + (block.as_ptr() as usize - addr);
     let mut i = 0;
 
     while i < len {
